Add a temporary way to use std::fstream instead of std::string for the Lexer
Lexer previously stored a string in the class. Now it's possible to use an fstream object instead
Add SourceProcessor class for the autotester to interact with the spa

Should the Lexer be a static class instead?
Problems with fstream is what if the file is not available during the lexification process? Need to add checks?